Don't return error when logs are discarded in awslogs
Docker logs an error every time the awslogs driver discards a message. Messages
are discarded when the internal awslogs buffer (channel) is full and the driver
is configured in non-blocking mode.
Logging every time a message is discarded can be problematic since Docker logs
can grow rapidly, potentially exhausting disk space.
Instead, it's preferable to log a warning periodically, displaying the number
of logs discarded for the period.

Signed-off-by: Angel Velazquez <angelcar@amazon.com>